Fix wrong rescaling in Heisenberg example
The problem became apparent when trying to work out a mutation example
in the context of the main theorem. After a lot of recalculations of
various constants it turned out that the mistake was in determining
gamma_3. After correcting this, somehow everything magically worked out
perfectly. This gives hope that the rescalings and normalizations
are correct now.